+++ scripts/tests/chipyaml/extensions/example_cluster.py
