+++ scripts/tests/chipyaml/extensions/wildcard_response_extractor_cluster.py
