+++ scripts/tests/chipyaml/paths_finder.py
